WIP commit for router/sockets feature
