+++ game/package.json
@@ -16,6 +16,7 @@
     "postgres": "^3.4.7",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.6.2",
     "vite-express": "^0.21.1"
   },
   "devDependencies": {
