+++ game/src/GameLobby.tsx
@@ -1,0 +1,12 @@
+import { useEffect, useMemo, useState } from 'react'
+import './App.css'
+import { TicTacToeClient } from './api'
+import { type Game, type indexes } from './game/game'
+
+
+
+export function GameLobby(){
+
+
+    
+}
